Add Trellis MoE GEMM kernel and associated tests
- Implemented the fused MoE GEMM kernel with Trellis 3-bit quantization in `gemm_trellis_moe.metal`.
- The kernel supports token routing, dequantization, SwiGLU activation, and expert probability weighting.
- Added a comprehensive test suite for the TrellisLinear layer in `test_trellis_linear.py`, covering decode and prefill scenarios, correctness against reference implementations, and memory efficiency.
- Included tests for various batch sizes and bias handling to ensure robustness of the implementation.

+++ metal_marlin/metal_dispatch.py
@@ -739,7 +739,7 @@ def mps_tensor_to_metal_buffer(
             raise RuntimeError("Failed to create Metal buffer for output tensor")
         return _CopyBackBuffer(buffer, tensor)
 
-    arr = tensor.detach().cpu().numpy()
+    arr = tensor.detach().cpu().float().numpy()
     buffer = device.newBufferWithBytes_length_options_(
         arr.tobytes(), arr.nbytes, Metal.MTLResourceStorageModeShared
     )
